requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
bcrypt==4.1.2
PyJWT==2.8.0
Pillow==10.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
tensorflow==2.15.0
scikit-learn==1.3.2
cryptography==41.0.7
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
psycopg2-binary==2.9.9
